+++ NEWS.md
@@ -3,7 +3,7 @@
 - MACT, NAICS, or SIC initially selected at launch of app now can be specified as parameters mact, naics, sic in ejamapp(), or as parameters default_mact, default_naics, default_sic in global_defaults_shiny.R, or in Advanced tab.
 - More consistent info about selected categories of sites now printed to console (and server log, depending on how app is hosted)
 - Fixed ejam2shapefile() where it had problems if closely related filename had previously been used
-- Fixed some edge cases related to invalid mact codes, too many points selected, etc.
+- Fixed server logic for validation of NAICS/SIC/MACT. Fixed some edge cases related to invalid mact codes, too many points selected, etc. Removed obsolete naics_validation() function. See better `naics_is.valid()`. 
 - Documented package dev-related utilities `grepn()` and `found_in_files()` (and also improved some internal/unexported utilities pkg_functions_* )
 
 # EJAM 2.32.7 (February 2026)